New Extractionsystem. Can unpack rar, 7z, zip, tar.gz, tar.bz2 The system is plugin able, to add more unpack routines easily. Has currently the same features as unrar, except deleting info files. I'll try to handle all other tickets regarding unrar or unpacking in gereral.
NOTE: There is currently a bug in the used lib with unpacking encrypted rar with encrypted filenames . I'll try to get in contact with the developer to solve the problem. Otherwise I have to port unrar as a plugin for the new system.
Currently for svn testing only. Will not be put in the nighlty until it's finished and the bug is solved.

git-svn-id: svn://svn.jdownloader.org/jdownloader/trunk@12847 ebf7c1c2-ba36-0410-9fe8-c592906822b4